+++ examples/typescript/core/src/app.ts
@@ -1,0 +1,9 @@
+import Pulse from '@pulsejs/core';
+
+const App = new Pulse();
+
+export default App;
+
+export const GlobalAppState = {
+  API_HOST_URL: App
+};