+++ examples/typescript/vue/src/App.vue
@@ -1,19 +1,34 @@
 <template>
   <div id="app">
-    <img alt="Vue logo" src="./assets/logo.png">
-    <HelloWorld msg="Welcome to Your Vue.js App"/>
+    <img alt="Vue logo" src="./assets/logo.png" />
+    <HelloWorld :msg="number" />
   </div>
 </template>
 
 <script>
-import HelloWorld from './components/HelloWorld.vue'
+import HelloWorld from './components/HelloWorld.vue';
+import Pulse from '@pulsejs/vue';
+
+const App = new Pulse();
+
+export const core = App.Core({
+  NUMBER: App.State('hello')
+});
 
 export default {
   name: 'App',
   components: {
     HelloWorld
+  },
+  created() {},
+  data() {
+    return {
+      ...this.mapCore(core => ({
+        number: core.NUMBER
+      }))
+    };
   }
-}
+};
 </script>
 
 <style>