+++ packages/pulse-react/lib/index.ts
@@ -4,6 +4,7 @@ import Pulse, { Integration } from '@pulsejs/core';
 export { PulseHOC } from './pulseHOC';
 export { usePulse } from './usePulse';
 export { useEvent } from './useEvent';
+export { useWatcher } from './useWatcher';
 
 export * from '@pulsejs/core';
 export default Pulse;